Mel responds in DMs
