+++ .gitignore
@@ -62,3 +62,4 @@ typings/
 
 # vim swapfiles
 *.swp
+ngrok
